+++ config/services.php
@@ -41,4 +41,9 @@ return [
         ],
     ],
 
+    'mailjet' => [
+        'key' => env('MAILJET_APIKEY'),
+        'secret' => env('MAILJET_APISECRET'),
+    ],
+
 ];