+++ database/migrations/2020_04_12_113223_add_tables_to_link_users_to_regions.php
@@ -1,0 +1,35 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AddTablesToLinkUsersToRegions extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('region_user', function (Blueprint $table) {
+            $table->integer('user_id');
+            $table->integer('region_id');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('region_user');
+
+        // Schema::table('region_user', function (Blueprint $table) {
+        //     //
+        // });
+    }
+}