+++ routes/web.php
@@ -22,13 +22,18 @@ Route::middleware(['auth'])->group(function () {
 
     //User
     Route::prefix('/users')->name('users.')->group(function () {
+        Route::get('/', 'UserController@index')->name('index')->middleware('can:do-admin');
+
         //Create
-        Route::get('/user/ajout', 'UserController@create')->name('create')->middleware('can:do-admin');
-        Route::post('/user', 'UserController@store')->name('store')->middleware('can:do-admin');
+        Route::get('/ajout', 'UserController@create')->name('create')->middleware('can:do-admin');
+        Route::post('/', 'UserController@store')->name('store')->middleware('can:do-admin');
 
         //Update
-        Route::get('/user/{user}', 'UserController@edit')->name('edit');
-        Route::put('/user/{user}', 'UserController@update')->name('update');
+        Route::get('/me', 'UserController@editSelf')->name('edit-self');
+        Route::put('/me', 'UserController@updateSelf')->name('update-self');
+        
+        Route::get('/{user}', 'UserController@edit')->name('edit')->middleware('can:do-admin');
+        Route::put('/{user}', 'UserController@update')->name('update')->middleware('can:do-admin');
     });
 });
 
@@ -52,7 +57,7 @@ Route::prefix('/approbations')->name('approvals.')->group(function () {
 //Places
 Route::prefix('/places')->name('places.')->group(function () {
     Route::middleware(['auth', 'can:do-moderation'])->group(function () {
-        // Route::get('/places', 'PlaceController@index')->name('index');
+        Route::get('/fermées', 'PlaceController@indexClosed')->name('closed');
       
         //Create
         Route::get('/ajout', 'PlaceController@create')->name('create');